ci: auto-sync local changes

+++ src/main/java/pro/datawiki/auth/base/repository/BalanceTransactionRepository.java
@@ -7,6 +7,6 @@ import pro.datawiki.auth.base.domain.BalanceTransaction;
 
 import java.util.List;
 
-public interface BalanceTransactionRepository extends JpaRepository<BalanceTransaction, Long> {
+public interface BalanceTransactionRepository extends JpaRepository<BalanceTransaction, Integer> {
     List<BalanceTransaction> findByUserIdOrderByCreatedAtDesc(Long userId);
 }
